refactoring and testing

+++ clean_repos.py
@@ -13,7 +13,6 @@ def clean_folder_recursive(path:str="repos", tqdm_active:bool=True):
   if not os.path.exists(path) or not os.path.isdir(path): return
 
   iterator = tqdm(os.walk(path)) if tqdm_active else os.walk(path)
-  unsuccess_deletes = []
 
   try:
     for dirpath, dirnames, filenames in iterator:
@@ -24,36 +23,42 @@ def clean_folder_recursive(path:str="repos", tqdm_active:bool=True):
 
       # Delete any blacklisted folder rightaway
       if any([path_ends_with(dirpath, end) for end in BLACKLIST_FOLDERS]):
-        try:
-          rmtree(dirpath)
-        except:
-          unsuccess_deletes.append(dirpath)
-        continue
+        if os.path.exists(dirpath):
+          success = False
+
+          while not success:
+            try:
+              rmtree(dirpath)
+              success = True
+            except:
+              time.sleep(0.1)
+              pass
+
+          continue
 
       for f in filenames:
         full_path = os.path.join(dirpath, f)
+        if not os.path.exists(full_path): continue
 
-        if not full_path.endswith("topics.txt") and not full_path.endswith(".py"):
+        if not full_path.endswith(".py"):
           if not path in full_path:
             print(f"Something went wrong\nInvalid path {full_path}")
             time.sleep(60)
             break
 
-          try:
-            os.remove(full_path)
-          except PermissionError:
-            os.chmod(full_path, 0o777)
-            os.remove(full_path)
-
-    for uns_path in unsuccess_deletes:
-      success = False
-      while not success:
-        try:
-          rmtree(dirpath)
-          success = True
-        except:
-          time.sleep(0.1)
-          pass
+          success = False
+          while not success:
+            try:
+              os.remove(full_path)
+              success = True
+            except PermissionError:
+              try:
+                os.chmod(full_path, 0o777)
+                os.remove(full_path)
+                success = True
+              except:
+                time.sleep(0.1)
+                pass
   except KeyboardInterrupt:
     pass
